% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{count}
\alias{count}
\title{Generate a data.frame with metrics computed from dataset
as columns and with additional column with current build id
columns are:
\itemize{
 \item{build}
 \item{lccfa}
 \item{lccpa}
 \item{lcpa}
 \item{lcfa}
 \item{rate}
 \item{sccfa}
 \item{sccpa}
 \item{scfa}
 \item{scpa}
 \item{tlfb}
 \item{tlpb}
 }}
\usage{
count()
}
\description{
Generate a data.frame with metrics computed from dataset
as columns and with additional column with current build id
columns are:
\itemize{
 \item{build}
 \item{lccfa}
 \item{lccpa}
 \item{lcpa}
 \item{lcfa}
 \item{rate}
 \item{sccfa}
 \item{sccpa}
 \item{scfa}
 \item{scpa}
 \item{tlfb}
 \item{tlpb}
 }
}

