% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{createDefaultQueryResult}
\alias{createDefaultQueryResult}
\title{Function selects data from customTravisTorrent table and returns it}
\usage{
createDefaultQueryResult(dbName, dbHost, dbLogin, dbPassword, ttTableName)
}
\arguments{
\item{dbName}{name of database}

\item{dbHost}{host of database}

\item{dbLogin}{user of database}

\item{dbPassword}{user's password to connect to database}

\item{ttTableName}{table from which you want to select}
}
\description{
Function selects data from customTravisTorrent table and returns it
}

