% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{insertToDB}
\alias{insertToDB}
\title{Function selects data from TravisTorrent table, creates new table (with additional column author_mail
with e-mail address of commit's author) and insert data into this new table. Values of author_mail column
are collected from Github using Github API}
\usage{
insertToDB(dbName, dbHost, dbLogin, dbPassword, clientId, clientSecret,
  newTableName, ttTableName, projects)
}
\arguments{
\item{dbName}{name of database}

\item{dbHost}{host of database}

\item{dbLogin}{user of database}

\item{dbPassword}{user's password to connect to database}

\item{clientId}{Github's client ID which is received when register to Github}

\item{clientSecret}{Github's client secret which is received when register to Github}

\item{newTableName}{name of a table which you want to create and insert data to}

\item{ttTableName}{name of TravisTorrent table which currently exists in your database}

\item{projects}{names of projects which you want to filter data table by, should be given
as string with comma as delimiter}
}
\description{
Function selects data from TravisTorrent table, creates new table (with additional column author_mail
with e-mail address of commit's author) and insert data into this new table. Values of author_mail column
are collected from Github using Github API
}

