% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committerMetrics.R
\name{lccpa}
\alias{lccpa}
\title{Compute the longest chain of past passed committer's builds}
\usage{
lccpa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the longest chain of past passed
committer's builds.
Past passed committer's builds are builds which:
\itemize{
 \item{ran on the same repository as current build,}
 \item{commit which triggered these builds was made
 by the same person who made commit which triggered current build,}
 \item{completed with status 'passed'}
}
}
\examples{
lccpa(123456)
}

