% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMetrics.R
\name{lcfa}
\alias{lcfa}
\title{Compute the longest chain of past failed builds}
\usage{
lcfa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the longest chain of past failed builds.
Past failed builds are builds which ran on the same repository
as current build and completed with status 'failed' or 'errored'
}
\examples{
lcfa(10297653)
}

