% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lcpa2}
\alias{lcpa2}
\title{Compute the squared error}
\usage{
lcpa2(currentBuild, qresult)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the elementwise squared error for a
number or a vector
}

