% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{queryResult}
\alias{queryResult}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 19396 rows and 10 variables:
\describe{
  \item{author_mail}{price, in US dollars}
  \item{gh_files_added}{price, in US dollars}
  \item{gh_files_deleted}{price, in US dollars}
  \item{gh_files_modified}{price, in US dollars}
  \item{gh_project_name}{price, in US dollars}
  \item{gh_src_churn}{price, in US dollars}
  \item{git_commit}{price, in US dollars}
  \item{tr_build_id}{price, in US dollars}
  \item{tr_started_at}{price, in US dollars}
  \item{tr_status}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
queryResult
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

