% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committerMetrics.R
\name{rate}
\alias{rate}
\title{Compute the ratio of passed build to all builds on repository}
\usage{
rate(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the ratio of passed builds to all builds
on repository. History of builds is taken from current build
}
\examples{
lccfa(123456)
}

