% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{replaceDatasetRDA}
\alias{replaceDatasetRDA}
\title{Function replaces dataset queryResult with another data, selected by user}
\usage{
replaceDatasetRDA(queryResult)
}
\arguments{
\item{queryResult}{new data}
}
\description{
Function replaces dataset queryResult with another data, selected by user
}

