% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMetrics.R
\name{scfa}
\alias{scfa}
\title{Compute the shortest chain of past failed builds}
\usage{
scfa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the shortest chain of past failed builds.
Past failed builds are builds which ran on the same repository
as current build and completed with status 'failed' or 'errored'
}
\examples{
scfa(10297653)
}

