% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{scpa}
\alias{scpa}
\title{Compute the shortest chain of passed builds}
\usage{
scpa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the shortest chain of passed builds
which ran on the same project as current build
before current build
}

