% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{selectFromRDA}
\alias{selectFromRDA}
\title{Function get data from dataset alldata.rda which contains all build from sample TravisTorrent table}
\usage{
selectFromRDA(clientId, clientSecret, projects)
}
\arguments{
\item{clientId}{Github's client ID which is received when register to Github}

\item{clientSecret}{Github's client secret which is received when register to Github}

\item{projects}{names of projects which you want to filter data table by, should be given
as string with comma as delimiter}
}
\description{
Function get data from dataset alldata.rda which contains all build from sample TravisTorrent table
}

