% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMetrics.R
\name{tlpb}
\alias{tlpb}
\title{Compute time from last passed build}
\usage{
tlpb(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the time (in minutes) elapsed from last
passed build (from the same repository as current build)
to time when current build have started
}

