% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{weightedRate}
\alias{weightedRate}
\title{Compute the squared error}
\usage{
weightedRate(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the elementwise squared error for a
number or a vector
}

