% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{count}
\alias{count}
\title{Function generates a data frame with metrics computed from dataset
as columns and with additional column with current build id.
Columns of data frame are presented below:
\itemize{
 \item{build}
 \item{lccfa}
 \item{lccpa}
 \item{lcpa}
 \item{lcfa}
 \item{rate}
 \item{sccfa}
 \item{sccpa}
 \item{scfa}
 \item{scpa}
 \item{tlfb}
 \item{tlpb}
 }}
\usage{
count(n = nrow(queryResult))
}
\arguments{
\item{n}{number of rows of queryResult}
}
\description{
Function generates a data frame with metrics computed from dataset
as columns and with additional column with current build id.
Columns of data frame are presented below:
\itemize{
 \item{build}
 \item{lccfa}
 \item{lccpa}
 \item{lcpa}
 \item{lcfa}
 \item{rate}
 \item{sccfa}
 \item{sccpa}
 \item{scfa}
 \item{scpa}
 \item{tlfb}
 \item{tlpb}
 }
}
\examples{
count(10)
}

