% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lccfa}
\alias{lccfa}
\title{Compute the squared error}
\usage{
lccfa(currentBuild, newData)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the elementwise squared error for a
number or a vector
}

