% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lcfa}
\alias{lcfa}
\title{Compute the longest chain of passed builds}
\usage{
lcfa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the longest chain of failed
and errored builds
which ran on the same project as current build
before current build
}

