% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMetrics.R
\name{lcpa}
\alias{lcpa}
\title{Compute the longest chain of past passed builds}
\usage{
lcpa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the longest chain of past passed builds.
Past passed builds are builds which ran on the same repository
as current build and completed with status 'passed'
}
\examples{
lcpa(10297653)
}

