% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lcpa}
\alias{lcpa}
\title{Compute the longest chain of past passed builds}
\usage{
lcpa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the shortest chain of past passed builds.
Past passed builds are builds which ran on the same repository
as current build
}
\examples{
lcpa(1763098)
}

