% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{queryResult}
\alias{queryResult}
\title{TravisTorrent builds data}
\format{A data frame with 19396 rows and 10 variables:

\describe{
  \item{tr_build_id}{id of build started at Travis-CI}
  \item{tr_status}{status of completed build}
  \item{author_mail}{e-mail address of person whose commit triggered build}
  \item{git_commit}{SHA of commit}
  \item{gh_project_name}{name of repository}
  \item{tr_started_at}{datetime when build has started}
  \item{gh_src_churn}{churn of source code on all commits included in build}
  \item{gh_files_added}{number of files added by all commits included in build}
  \item{gh_files_modified}{number of files modified by all commits included in build}
  \item{gh_files_deleted}{number of files deleted by all commits included in build}
}}
\usage{
queryResult
}
\description{
A dataset containing data relating to builds which ran on Travis-CI
}
\examples{
head(queryResult)
}
\keyword{datasets}

