% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{queryResult}
\alias{queryResult}
\title{TravisTorrent builds data}
\format{A data frame with 19396 rows and 10 variables:
\describe{
  \item{tr_build_id}{price, in US dollars}
  \item{tr_status}{weight of the diamond, in carats}
  \item{author_mail}{weight of the diamond, in carats}
  \item{git_commit}{weight of the diamond, in carats}
  \item{gh_project_name}{weight of the diamond, in carats}
  \item{tr_started_at}{weight of the diamond, in carats}
  \item{gh_src_churn}{weight of the diamond, in carats}
  \item{gh_files_added}{weight of the diamond, in carats}
  \item{gh_files_modified}{weight of the diamond, in carats}
  \item{gh_files_deleted}{weight of the diamond, in carats}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
queryResult
}
\description{
A dataset containing data relating to builds which ran on Travis-CI
}
\examples{
head(queryResult)
}
\keyword{datasets}

