% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMetrics.R
\name{scpa}
\alias{scpa}
\title{Compute the shortest chain of past passed builds}
\usage{
scpa(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the shortest chain of past passed builds.
Past passed builds are builds which ran on the same repository
as current build and completed with status 'passed'
}
\examples{
scpa(10297653)
}

