% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMetrics.R
\name{tlfb}
\alias{tlfb}
\title{Compute time from last failed build}
\usage{
tlfb(currentBuild)
}
\arguments{
\item{currentBuild}{id of travis-ci build}
}
\description{
This function computes the time (in minutes) elapsed from last
failed build (from the same repository as current build)
to time when current build have started
}

